tests(request_parser: implemented first tests for header parsing)

+++ Includes/request_parser/RequestParser.hpp
@@ -79,7 +79,7 @@ private:
 	bool	finished_headers = false;
 	bool	finished = false;
 
-	long					pos;
+	size_t					pos;
 	const static std::regex	method_pat;
 	const static std::regex	request_line_pat;
 	const static std::regex	uri_pat;
